Tags tests: don't assign back to testCase reference in test loop (again)
Follow-up the last commit (see its commit message) and fix another
assignment to testCase, this one was added in 2ad030438 (Implement
UserCreateTag RPC, 2017-09-26). While we're at it let's be consistent
here with later tests I've added and test the entire response, not
just response.Tag.

That re-alignment of "[]byte(inputTagName)" is "make format" being
stupid, why not align all the key-values just because there's a tiny
comment in the middle, but I guess it's something that makes sense for
larger comments...