hooks: Derive GL_REPOSITORY from passed in repo
One of the GL_ values is GL_REPOSITORY, which is used both for
authentication against GitLab's API as well as for the custom hooks
environment. It's currently passed as an explicit variable when invoking
hooks, but this is in fact not at all necessary as we already have the
`gitalypb.Repo`.

Simplify the code to not expect GL_REPOSITORY as an explicit variable
anymore, instead deriving it from the Repo protobuf.