hooks: Discern expected and actual environment for custom hook tests
Right now, our custom hook tests expect the same environment for the
invoked custom hooks as the environment that we provide. This is going
to change though as we move more variables into the hooks payload. As
such, the input environment is only going to contain the payload, while
the custom hooks environment will stay the same.

Let's prepare for this change by converting tests to discern between
actual and expected environment to make it easier to change.