Support cloning with an object pool in CreateFork
This is needed to support fast forking. When an object pool is provided,
forking can cheap in terms of disk space and time since the clone only
needs to fetch the references.

The `git clone` can take in a `--reference <repository>` flag and will
output an alternates file with absolute paths. To maintain the use of
relative paths, after the fork is successful we recreate the alternates
file.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/24523