hooks: Introduce ReceiveHooksPayload
Next to the values we already inject via the HooksPayload, there's also
some information which only applies to the git-receive-pack(1) hooks
pre-receive, update and post-receive. This includes information about
the user and about the protocol which is used for the push, both which
are used to establish whether the current push is allowed to proceed.
This commit thus introduces a new optional ReceiveHooksPayload structure
to the HooksPayload, which contains this information.

As with the other fields, we also need to have fallback code in place to
enable zero-downtime upgrades where the server is still running the old
code, while the gitaly-hooks executable was already updated. We thus
have a fallback in place to read the GL_PROTOCOL, GL_ID and GL_USERNAME
values in case no payload was provided.