git: Disambiguate global and command options
Currently, we use the Option type both for global and command-specific
options. This use is confusing though as it is possible to pass
non-global options as global ones and vice versa. Two examples exist
already where the interface is lacking with the `ConfigPair` and
`SubSubCmd` types, which both make no sense at all as a global option
and would break the git command.

Let's disambiguate both usages by introducing a new `GlobalOption` type.
While it has the same implementation for most options, above examples
don't implement it. As a result, they cannot be used in global context
anymore.