Clean up params usage for the methods of the Coordinator
The grpcCall struct was introduced in the c954a53a1 (Praefect:
proper multi-virtual storage support, 2020-04-28) commit
to reduce the number of parameters used in methods. It contains
all info needed to execute proxy operation and there is no need
to pass target repository or anything like that as a separate
parameter because those are already present in the grpcCall.

This commit removes redundant parameters from the methods and
substitutes their usage with usage of the grpcCall parameter
that has the same value.