updateref: Add option to disable reference-transactions
The updateref package will currently alwyas set up the Updater in such a
way that it uses the reference-transaction hook. This may not be wanted
in some contexts, which is why this commit now sets up a new option to
disable hooks altogether.

The way Options are implemented is really quite simplistic and extremely
limited. It's likely we won't ever add new options though, which makes
it fit for the current task at hand but for nothing else. It should be
easy enough to extend though, and using options is preferred to
introducing a new required parameter as the latter would require us to
change all callers.