remote: Disable transactions for FetchInternalRemote
With the same reasoning as the preceding commit, this disables the
reference-transaction hook for `FetchInternalRemote`.

Note that the test change to create a single commit in the source
repository is sufficient to demonstrate that hooks aren't executed
anymore. Previously, there simply weren't any changes which were
fetched, which caused us to invoke no hooks at all. With this change, we
now create a commit and would thus update a reference in the target
repository, but as the test setup was incomplete and didn't set up the
hooks service at all, it would've failed to execute the hook now. But as
we now disabled hooks, it continues to succeed.