objectpool: Disable transactions for FetchIntoObjectPool
When executing git-fetch(1), git will currently create one referecne
transaction per reference that is to be modified. This is quite
inefficient when used together with the reference-transaction hook, as
it effectively means that we execute the hook once for each reference.
If fetching hundreds of thousands of references, this directly
translates to hundreds of thousands of spawned processes, which means an
extreme performance hit.

Fix the issue for `FetchIntoObjectPool` by disabling the
reference-transaction hook for now. This is only a stop-gap measure
until we land the equivalent of `git fetch --atomic`, which creates a
single transaction only.