Tags tests: test fewer nesting levels
Amend test code I added in 5896fdaf4 (UserCreateTag tests: test nested
annotated tags, 2020-12-08) to be less expensive. As pointed out in
[1] we're not getting anything extra from testing 10 nested levels,
and spending a lot of time doing so.

Also change the loop test from "<" to "<=" so the nested level refers
to the level of nested tags. I.e. tag->tag->commit is just 1 level of
nesting, not 2 or 3.

1. https://gitlab.com/gitlab-org/gitaly/-/merge_requests/2905#note_465020107