Handle empty notification messages
The triggers may be called with no actual change done by the
operation, like updating of the non-existing rows. In this case
existing trigger will generate a notification with empty payload.
This payload can't be properly handled by the listener and will
cause cache disabling.

To deal with it we add check in the trigger logic to
verify if there are any actual changes done to the state.
This change also includes structural changes into the query but
preserve the same logic and format of the payload.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/3309