Makefile: Remove unused argument to `find_go_sources`
In order to find all Go packages, we piggy-back on `find_go_sources` and
simply use their respective directories as value. We used to do this via
a call to sed(1), but have since replaced this with the `dir` builtin
function. The conversion forgot to remove the script argument for sed
though.

Remove the unused argument to avoid any confusion.