Makefile: Deduplicate Go packages
In order to find all Go packages, we currently piggy-back on
`find_go_sources` and use their respective directories. What wasn't
considered though is that this causes us to list each directory as many
times as there are `*.go` files in it. It doesn't hurt us right now as
Go seemingly deduplicates the list, but it is hard to parse in verbose
mode and may easily cause us to hit command line size limits.

Fix the issue by deduplicating the directories.