hooks: Refactor PreReceiveHook to explicitly receive push options
While the `PostReceiveHook` has an explicit parameter for push options,
the `PreReceiveHook` doesn't. They both have the same kind of code paths
though, and passing push options as an explicit parameter allows us to
unify setup of the custom hooks directory for both hooks in a later
patch.

Refactor the `PreReceiveHook` function to receive an explicit slice of
push options to allow for this unification.