hooks: Move `GitPushOptions` into the hook manager's code
The `GitPushOptions` function is used to convert a slice of push options
into a set of environment variables like git would export them. Except
for tests, this code is really only used in the hooks manager's code, so
let's move it there in order to have computation of all custom hook
environment variables in a single place.