hooks: Stop testing for GL_REPO environment variable
Some of our tests are verifying that an environment variable GL_REPO is
being passed through to custom hooks, but in fact this variable should
never be passed through. The repository is only ever passed through with
GL_REPOSITORY, and GL_REPO isn't documented anywhere.

Remove the checks to avoid any confusion and to make way for improved
filtering of custom hooks environment variables in a subsequent patch.