Revert "Merge branch 'po-user-update-submodule-ff-default' into 'master'"
This reverts commit 61c631b3e50b2a4db5c6c821d1d22906260f2f28, reversing
changes made to 64d6e3dc24637c9755c5ca5220c1330d260689e5.

Because of the bug noted in my
https://gitlab.com/gitlab-org/gitaly/-/merge_requests/2926 of not
ferrying errors upwards. The feature has been stopped already, my MR
fixes it, but as discussed in Slack we'd like to have master in a
known-good state for an upcoming release.