command: Filter `AllowedEnviroment()` via explicit slice
The `AllowedEnvironment()` function currently operates on the real
environment directly, which makes it quite inflexible as one cannot
filter a different environment. As we're about to move filtering of the
allowed environment into the hooks manager, which only has those
variables in the incoming environment slice, this commit refactors the
function to instead filter a given slice.