hooks: Filter custom hooks commands in the hooks manager
We're currently filtering environment variables only in the gitaly-hooks
command. This is a bit awkward, as we need to pass on variables to the
hooks service which is not intended for custom hooks. One such example
is the hooks payload, which we need to pass on but which shan't be
passed to the custom hooks.

This commit thus starts filtering environment variables in the hooks
manager. This fixes a known bug where we pass the hooks payload to
custom hooks. Filtering in the gitaly-hooks binary is currently still in
place, but will be removed in the subsequent commit.