hooks: Remove test setups for gitlab-shell configuration
Our hooks used to be implemented via gitlab-shell, but all production
code relating to this is gone by now as we have moved to our own
gitaly-hooks executable. There is still testhelpers related to the
gitlab-shell configuration file though, which muddy the water as they
are seemingly required while in fact they're not used at all.

Remove all code related to the gitlab-shell configuration to clean this
up. Note that the gitlab-shell secret is still required and thus not
removed, as it is used to authenticate against the GitLab API.