smarthttp: Remove useless hookdir overrides
The smarthttp tests set up several overrides for the hook directory,
which can be helpful to execute a different set of hook binaries which
the caller supplies. In some cases, we set it to the empty value or to
the path of the `ruby/git-hooks` directory. Those aren't required at all
as the empty value is the default, and if the default is in place then
we'll figure out the `ruby/git-hooks` directory on our own.

Remove those overrides to clean up the tests.