gitaly-hooks: Remove filtering of environment variables
Now that we have started to filter environment variables in the hooks
manager, we can just pass on all variables which gitaly-hooks has set to
the hooks service. This removes logic duplication, simplifies logic in
gitaly-hooks and also allows us to easier extend our use of environment
variables in the hooks context.