hooks: Pass GIT_DIR to custom hooks
Back with the Ruby implementation of hooks, we used to pass the GIT_DIR
environment variable to custom hooks, which allows any spawned git
executables to corretcly locate the repository in which it is to run.
This variable got dropped when we converted the code from Ruby to Go.

This commit adds it back in to fix any custom hooks which rely on it
being set correctly.