Feature flag rollout doc: rewrite & make better use of template
Rather than duplicating things between the issue template and process
doc, let's cross link the two.

Also elaborate on a lot of things I encountered when enabling my first
features. This is the result of feedback on the initial version of
this [1] and a video chat with @zj-gitlab this morning.

1. https://gitlab.com/gitlab-org/gitaly/-/merge_requests/2910