Break dependency of the catfile on global config.Config
The catfile package directly or indirectly uses config.Config
value. And as we are about to remove config.Config from the
code we need to break this dependency and substitute with
other abstractions.

For now those are storage.Locator and alternates.Env instead
of the old alternates.PathAndEnv that depends on the global var.

It was decided to extend constructor function of the catfile
to accept storage.Locator as a dependency as it is already
injected in most of the "server"s which handle requests.
The other option could be creation of the "factory", but it will
require a new dependency to be injected into all existing
services which is a lot more changes and adds currently unnecessary
abstraction that needs to be managed.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699