git: Extend SafeCmd tests to test GlobalOption
While the Options interface has its fair share of tests, GlobalOptions
don't yet. This is mostly as it has only been introduced fairly recently
and because its implementation was the same as that for Options anyway.
So adding tests back when the GlobalOptions interface was created didn't
make a lot of tests.

This is about to change though as we're going to add the first option
which has different behaviour based on whether it's invoked as Option or
GlobalOption. So let's prepare for this by adding some testcases for our
existing GlobalOption implementations.