golangci: Enable missing default linters
The golangci-lint project has a list of linters [1] which are enabled by
default, which indicates that they're deemed useful enough to be always
enabled. As we're default-disabling all linters, we do not have those
enabled for Gitaly.

Given that they don't add any warnings, let's enable all defaul linters.
The only exception is the errcheck linter as we have quite some
locations where we don't perform error checking at all.

[1]: https://golangci-lint.run/usage/linters/