ref: Preallocate arrays to avoid needless reallocations
When computing which references we are about to delete in the DeleteRefs
RPC, we allocate two arrays into which we put the final reference names.
We're doing this quite inefficiently though by just calling `append`,
even though we know the size of both arrays beforehand.

Improve our code to instead preallocate arrays to avoid repeated
reallocations.