ref: Use repository abstraction to list existing references
When determining which references to delete for the DeleteRefs RPC, we
shell out to git-for-each-ref(1) to determine which references match a
given set of prefixes. We do have a higher-level abstraction around to
list references via git-for-each-ref(1) though in our LocalRepository
interface.

Convert the code to use `LocalReposiory.GetReferences()` to have one
less moving part where we manually invoke git.