UserCreateTag tests: add a lightweight tag to our "tag" object testing
Improve the tests I added in 5896fdaf4 (UserCreateTag tests: test
nested annotated tags, 2020-12-08) to, for each level of our tag
object nest, also create a "lightweight" tag to our just-created "tag"
object.

Of course there's no such thing in git, we'd just create a new refname
for a "tag" object whose name doesn't match the envelope. Under fsck
--strict this is a non-fatal "invalid 'tag' name" warning.

But as seen here what the UserCreateTag API does is something
completely bizarre. It sees "oh, a lightweight tag", and just uses
"^{}" somewhere to create a normal lightweight tag to our eventual
target, completely ignoring the "tag" object nest in-between.

The Ruby encoding blurb is needed because otherwise we'll have e.g.:

    pre-receive:15:in `<main>': invalid byte sequence in US-ASCII (ArgumentError)

This is because a lot of these tags are creating a tag called
"to-be-créated-soon".