handle start and target repositoires being the same in UserCommitFiles
Gitaly receives requests in UserCommitFiles where the start and the
target repositories are the same. In such cases, it doesn't have to
resolve the branches from start repository as it is the same. This
commit checks if the repositories are the same and only operates on
the local repository if that is the case.