auth: Introduce function to set token validity duration
Right now, it's not possible to modify the duration for which a token is
valid outside of the `auth` package, but in fact we do rely on being
able to do so in the authentication tests in `internal/gitaly/server`.
As it's not possible to modify it via code, we instead rely on modifying
its value via the Go's `-X` linker flag, which is quite awkward.

This commit introduces a new function `SetTokenValidityDuration()` which
allows us to fix this weirdness.