token: Remove `tokenValidity` initialization
In order to support our ldflags hack which was removed in the preceding
commit, we had separate variables for the token validation string and
the actual duration. This was done so that one could pass a `-X` flag to
the linker in order to override the string value and parse it into the
real duration at initialization time. But given that we don't use that
hack anymore, it's now useless.

Let's remove this infrastructure now and just use the timeout directly.