Removal of config.Config usage from objectpool
The objectpool package depends on the global config.Config
which makes testing harder and leads to races as well.

This change resolves this dependency by adding config.Cfg
as an external dependency to the constructor of the objectpool
implementation throw the NewServer function. And replaces
all usages of the global var with the bound to the server entity.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699