Replacement of config.NewLocator with existing dependency
The objectpool package uses call config.NewLocator throwout
the code without need as this dependency already included
as a field to the server type.

The change replaces usage of the config.NewLocator with existing
instances, so we are not dependent on exact implementation
here anymore.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699