stats: Fix stop condition when parsing packfile negotiations
When the packfile negotiation parser sees the "done" pktline, then it
should stop parsing as everything that follows is the actual packfile's
contents. We do so, or at least we intend to do so, but in fact the
`break` we have there is ineffective as it breaks the case statement,
not the loop.

Fix this by correctly breaking out of the loop if we've seen a "done"
pktline.