golangci: Enable error and documentation linters
This commit enables a set of linters which the project currently doesn't
abide to in a lot of places:

    - errcheck verifies that error returns are checked correctly.

    - errorlint verifies that Go 1.13-style errors are handled correctly
      by using the new `%w` formatter to wrap errors and using
      `errors.Is()` to check them.

    - golint verifies various things, but the most important one is that
      it checks that each exported symbol has proper documentation.

Fixing all errors pointed out by those linters at once is infeasible, as
that'd require hundreds of changes altogether. But the reported problems
are useful regardless of that and may point out real errors and style
violations.

So let's enable those linters for now, but only for new changes. While
this in fact influences all linters, there are currently no problems
reported previous to this commit. So effectively, it doesn't make a
difference for those other linters.