User{Create,Delete}Tag tests: add tests for invalid tag names
In a later migration of UserCreateTag from Ruby to Go we'll want to
make sure that we're not vulnerable to tag injections as we construct
a "git mktag" tag signature.

I'm also adding some tests for "UserDeleteTag" here for good measure,
although in that case mainly to test the consistency of the error
messages. I don't think we'll have anything like the potential mktag
problem there.