golangci: fix linting issues on "master"
I broke "make lint" on master in my a2b32a6b8 (Merge branch
'avar/port-user-tag-create-to-go-prep-2' into 'master', 2020-12-23).

I thought that the function of "make lint" was to check the latest
commit, and couldn't make sense of it warning about just 2 occurrences
of code in tags_test.go out of >10 uses of the exact same code.

It turns out this is because of 20ef26439 (golangci: Enable error and
documentation linters, 2020-12-17) which started warning about this
since the commit noted in that commit.

I think we should probably go for something like 0b5a8ea8e (disable
gochecknoglobals linter, 2020-12-17) instead. We're going to start
warning about all uses of that testhelper, but in any case, let's just
narrowly fix the smoke on master now.